% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{addDataSource}
\alias{addDataSource}
\title{Append a well-formatted data_source record to the backbone.data_source table in PostGIS}
\usage{
addDataSource(connectionDetails, dataSource)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{dataSource}{(data.frame) A well-formatted data source record. Typically created using gaiaSourceCreator RShiny.}
}
\value{
A new record in the backbone.data_source table in PostGIS
}
\description{
Append a well-formatted data_source record to the backbone.data_source table in PostGIS
}
