% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndices.R
\name{appendGeomIndexRecord}
\alias{appendGeomIndexRecord}
\title{Append a geom_index record to the geomIndexTable object}
\usage{
appendGeomIndexRecord(geomIndexTable, dataSourceRecord)
}
\arguments{
\item{geomIndexTable}{(data.frame) the existing geomIndexTable}

\item{dataSourceRecord}{(data.frame) A full record (entire row) from the backbone.data_source table}
}
\value{
(data.frame) updated geomIndexTable
}
\description{
Append a geom_index record to the geomIndexTable object
}
