% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assignGeomIdToAttr}
\alias{assignGeomIdToAttr}
\title{Join a column of geom_X record identifiers to a staged attr_X table}
\usage{
assignGeomIdToAttr(connectionDetails, stagedResult, geomIndexId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{stagedResult}{(data.frame) A table standardized in the attr_template or geom_template mold}

\item{geomIndexId}{(integer) Identifier of a record in the backbone.geom_index table. Usually sourced from the \code{attr_of_geom_index_id} entry of an attr_index record}
}
\value{
(data.frame) An updated \code{stagedResult} table with \code{geom_record_id}s corresponding to a geom_X table appended
}
\description{
Join a column of geom_X record identifiers to a staged attr_X table
}
