% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{checkTableExists}
\alias{checkTableExists}
\title{Check if a table exists in a database}
\usage{
checkTableExists(connectionDetails, databaseSchema, tableName)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{databaseSchema}{(character) schema that contains the table to be checked}

\item{tableName}{(character) name of the table to be checked}
}
\value{
(boolean) A logical value indicating whether the table exists
}
\description{
Check if a table exists in a database
}
