% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{checkVariableExists}
\alias{checkVariableExists}
\title{Check if a variable exists in an attr_X table}
\usage{
checkVariableExists(connectionDetails, databaseSchema, tableName, variableName)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{databaseSchema}{(character) schema that contains the table containing the variable to be checked}

\item{tableName}{(character) name of the table containing the variable to be checked}

\item{variableName}{(character) name of the variable to be checked}
}
\value{
(boolean) A logical value indicating whether the variable exists
}
\description{
Check if a variable exists in an attr_X table
}
