% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connectGaia}
\alias{connectGaia}
\title{Connect to a Gaia database}
\usage{
connectGaia(connectionDetails)
}
\arguments{
\item{connectionDetails}{Output of \code{\link[=createGaiaConnectionDetails]{createGaiaConnectionDetails()}}.}
}
\value{
A DatabaseConnector connection.
}
\description{
Connect to a Gaia database
}
