% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{createAttrIndexRecord}
\alias{createAttrIndexRecord}
\title{Create a single record in the backbone.attr_index table}
\usage{
createAttrIndexRecord(connectionDetails, dataSourceRecord)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{dataSourceRecord}{(data.frame) A full record (entire row) from the backbone.data_source table}
}
\value{
A new record in the backbone.attr_index table
}
\description{
Create a single record in the backbone.attr_index table
}
\examples{
\dontrun{

record <- getDataSourceRecord(connectionDetails = connectionDetails, dataSourceUuid = 9999)

createAttrIndexRecord(connectionDetails = connectionDetails, dataSourceRecord = record)
}

}
