% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndices.R
\name{createIndices}
\alias{createIndices}
\title{Create and populate backbone.geom_index and backbone.attr_index from registered sources in backbone.data_source}
\usage{
createIndices(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
Two tables (backbone.geom_index and backbone.attr_index) in PostGIS
}
\description{
Create and populate backbone.geom_index and backbone.attr_index from registered sources in backbone.data_source
}
\examples{
\dontrun{
createIndices(connectionDetails = connectionDetails)
}

}
