% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createSchemaString}
\alias{createSchemaString}
\title{Create the schema and sanitize source values}
\usage{
createSchemaString(rec)
}
\arguments{
\item{rec}{(data.frame) A full record (entire row) from the backbone.data_source table}
}
\value{
(character) Sanitized schema string
}
\description{
Create the schema and sanitize source values
}
