% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{disconnectGaia}
\alias{disconnectGaia}
\title{Disconnect from a Gaia database}
\usage{
disconnectGaia(connection)
}
\arguments{
\item{connection}{A connection from \code{\link[=connectGaia]{connectGaia()}}.}
}
\description{
Disconnect from a Gaia database
}
