% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocodeAddresses}
\alias{geocodeAddresses}
\title{Geocode a table with addresses}
\usage{
geocodeAddresses(addressTable)
}
\arguments{
\item{addressTable}{(data.frame) A table with a column of addresses prepared
for the Degauss geocoder. Addresses must be
formatted such as "123 Main Street Boston MA 02108"
See https://degauss.org/using_degauss.html#Geocoding
for more detailed information on input address
formatting}
}
\value{
(data.frame) The original input table with columns lat and lon
instead of the address column
}
\description{
Geocode a table with addresses
}
\examples{
dontrun
{
  geocodeAddresses(addressTable = cohortWithAddresses)
}

}
