% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getAttrIndexTable}
\alias{getAttrIndexTable}
\title{Get the backbone.attr_index table}
\usage{
getAttrIndexTable(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) The entire backbone.attr_index table
}
\description{
Get the backbone.attr_index table
}
