% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getCohortAddresses}
\alias{getCohortAddresses}
\title{Get Cohort Addresses}
\usage{
getCohortAddresses(connectionDetails, cdmDatabaseSchema, cohort)
}
\arguments{
\item{connectionDetails}{(list) For connecting to an OMOP server. An object of class connectionDetails as created by the createConnectionDetails function}

\item{cohort}{(data.frame) A table that includes a columns named "SUBJECT_ID" that contains a list of OMOP person_ids. Ideally, an OHDSI cohort table}

\item{cdmDatabseSchema}{(character) Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.}
}
\value{
(data.frame) An augmented Cohort table with patient addresses attached
}
\description{
Get Cohort Addresses
}
\examples{
\dontrun{
getCohortAddresses(connectionDetails = connectionDetails,
                   cdmDatabaseSchema = "myDatabase.dbo",
                   cohort = cohort)
}

}
