% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getDataSourceTable}
\alias{getDataSourceTable}
\title{Get the entire backbone.data_source table}
\usage{
getDataSourceTable(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) The entire backbone.data_source table
}
\description{
Get the entire backbone.data_source table
}
