% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getGeomIndexRecord}
\alias{getGeomIndexRecord}
\title{Get a record from the backbone.geom_index table from it's corresponding UUID in backbone.data_source table}
\usage{
getGeomIndexRecord(connectionDetails, dataSourceUuid)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{dataSourceUuid}{(UUID) The UUID for the data source that is registered in the backbone.data_source table}
}
\value{
(data.frame) A full record (entire row) from the backbone.geom_index table
}
\description{
Get a record from the backbone.geom_index table from it's corresponding UUID in backbone.data_source table
}
