% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getGeomTemplate}
\alias{getGeomTemplate}
\title{Get the geom_template table}
\usage{
getGeomTemplate(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) An empty geom_template table
}
\description{
Get the geom_template table
}
