% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getLocationAddresses}
\alias{getLocationAddresses}
\title{Get Location Table Addresses}
\usage{
getLocationAddresses(connectionDetails, cdmDatabaseSchema)
}
\arguments{
\item{connectionDetails}{(list) For connecting to an OMOP server. An object of class connectionDetails as created by the createConnectionDetails function}

\item{cdmDatabseSchema}{(character) Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.}
}
\value{
(data.frame) A transformed Location table with addresses and any existing latitude and longitudes attached
}
\description{
Get Location Table Addresses
}
\examples{
\dontrun{
getLocationAddresses(connectionDetails = connectionDetails,
                   cdmDatabaseSchema = myDatabase.dbo)
}

}
