% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{getStaged}
\alias{getStaged}
\title{Download and import data from a web-hosted source}
\usage{
getStaged(rec)
}
\arguments{
\item{rec}{(data.frame) A full record (entire row) from the backbone.data_source table corresponding to the data source of interest. Usually created using \code{getDataSourceRecord} function}
}
\value{
(data.frame) An untransformed version of the source data
}
\description{
Download and import data from a web-hosted source
}
\examples{

\dontrun{
staged <- getStaged(dataSourceRecord)
}

}
