% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{importGeomIndexTable}
\alias{importGeomIndexTable}
\title{Import the geom_index table to PostGIS}
\usage{
importGeomIndexTable(connectionDetails, geomIndexTable)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{geomIndexTable}{(data.frame) The entire backbone.geom_index table}
}
\value{
A populated backbone.geom_index table in PostGIS
}
\description{
Import the geom_index table to PostGIS
}
