% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGeometry.R
\name{importGeomTable}
\alias{importGeomTable}
\title{Import a well-formatted geometry table into an empty instance of geom_X in PostGIS}
\usage{
importGeomTable(connectionDetails, staged, geomIndex)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{staged}{(data.frame) A well-formatted geometry table. Created by appending staging data table to an geom_template}

\item{geomIndex}{(data.frame) A full record (entire row) from the backbone.geom_index table corresponding to the registered geometry of interest}
}
\value{
A table (geom_X) in PostGIS
}
\description{
Import a well-formatted geometry table into an empty instance of geom_X in PostGIS
}
