% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{listDatasources}
\alias{listDatasources}
\title{Datasets registered in the database}
\usage{
listDatasources(connection)
}
\arguments{
\item{connection}{A connection from \code{\link[=connectGaia]{connectGaia()}}.}
}
\value{
A data frame with the registered datasets (\code{backbone.data_source}).
}
\description{
Datasets registered in the database
}
