% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{readFromTar}
\alias{readFromTar}
\title{Unrar and read contents of a rar file into R memory}
\usage{
readFromTar(tarfile, exdir, rec)
}
\arguments{
\item{exdir}{(character) path to where contents of the compressed file should be extracted}

\item{rec}{(data.frame) A full record (entire row) from the backbone.data_source table corresponding to the data source of interest. Usually created using \code{getDataSourceRecord} function}

\item{rarfile}{(character) path to the compressed file}
}
\value{
(data.frame) An untransformed version of the source data
}
\description{
Unrar and read contents of a rar file into R memory
}
