% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_offline_storage}
\alias{setup_offline_storage}
\title{Configure offline storage}
\usage{
setup_offline_storage(dir = "/opt/data", persist_data = FALSE)
}
\arguments{
\item{dir}{(string) The directory to be used for offline storage. Default is "/opt/data".}

\item{persist_data}{(boolean) If TRUE, the downloaded files will persist indefinitely. If FALSE, they will not persist after the session ends. Default is FALSE.}
}
\value{
Copies the storage.yml file to the working directory and configures it with the specified directory and persistence settings. Alternatively, the file can be edited directly.
}
\description{
Configure offline storage
}
\examples{
\dontrun{

setup_offline_storage(dir = "/path/to/data", persist_data = TRUE)
}

}
