% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{standardizeStaged}
\alias{standardizeStaged}
\title{Create a standardized version of the staged data that was imported from source}
\usage{
standardizeStaged(staged, spec)
}
\arguments{
\item{staged}{(data.frame) Table of attributes that was imported from a source by \code{getStaged} function}

\item{spec}{(JSON) The attr_spec from the backbone.variable_source table or the geom_spec from backbone.data_source.}
}
\value{
(data.frame) A table standardized in the attr_template or geom_template mold
}
\description{
Create a standardized version of the staged data that was imported from source
}
\examples{

\dontrun{
stagedResult <- standardizeStaged(staged = staged, spec = variableTable$attr_spec)
}

}
